backend/requirements.txt
ccxt==4.3.68
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
websockets==13.0.1
asyncio==3.4.3
aiohttp==3.10.5
python-multipart==0.0.9
